@@ -12,12 +12,13 @@
       You also get clean error messages with a reference to the line of code that faulted.
     </Description>
     <RepositoryUrl>https://github.com/holonsoft/FluentConditions</RepositoryUrl>
-    <Version>1.6.1</Version>
-    <TargetFrameworks>netcoreapp3.1;net5.0;netstandard2.0;net472;net48;net6.0</TargetFrameworks>
+    <Version>1.6.2</Version>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;net6.0</TargetFrameworks>
     <LangVersion>latest</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <ImplicitUsings>enable</ImplicitUsings>
     <RepositoryType>GIT</RepositoryType>
-    <PackageTags>#fluent#conditions#net5#netcore3.1#netcoreapp3.1#net472#net48#net6</PackageTags>
+    <PackageTags>fluent;conditions;netstandard2.1;netstandard2.0;net6</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
