@@ -1,21 +1,19 @@
-﻿namespace holonsoft.FluentConditions
+﻿namespace holonsoft.FluentConditions;
+public readonly ref struct ConditionValueHolder<T>
 {
-	public readonly ref struct ConditionValueHolder<T>
-	{
-		internal readonly T _value;
+  internal readonly T Value;
 
-		internal readonly string _valueName;
-		internal readonly string _callerMemberName;
-		internal readonly string _sourceFilePath;
-		internal readonly int _sourceLineNumber;
+  internal readonly string ValueName;
+  internal readonly string CallerMemberName;
+  internal readonly string SourceFilePath;
+  internal readonly int SourceLineNumber;
 
-		public ConditionValueHolder(T value, string valueName, string callerMemberName, string sourceFilePath, int sourceLineNumber)
-		{
-			_value = value;
-			_valueName = valueName;
-			_callerMemberName = callerMemberName;
-			_sourceFilePath = sourceFilePath;
-			_sourceLineNumber = sourceLineNumber;
-		}
-	}
+  public ConditionValueHolder(T value, string valueName, string callerMemberName, string sourceFilePath, int sourceLineNumber)
+  {
+    Value = value;
+    ValueName = valueName;
+    CallerMemberName = callerMemberName;
+    SourceFilePath = sourceFilePath;
+    SourceLineNumber = sourceLineNumber;
+  }
 }
